[http] Introduce specialised `HeaderValue` instead of plain `String`s
This makes transforming the response easier for shared code paths, for
example when the header is used in multiple places, this avoids
serializing and deserializing which improves performance.

`HeaderValue` can also be used to restrict the types of setting various
`HeaderName`s. Most headers have a strict format, making them less
error-prone for the handler.

Another advantage is that we can have a simpler API to use, such as the
`SecFetchDest` enum.

At last, this removest the deserialization strain from the handler to
the transport code.