[quality] Decorate functions with `#[must_use]` where needed
This isn't really a problem, since the APIs are designed where the
return type is always needed, but for correctness this is still a good
thing to have.

It indicates which functions yield a result that is always needed to be
used/handled. Functions without this value might return e.g. a size read
from I/O, where the size can also be inferred from the `String` that was
read into.