[cache] Avoid caching racing
Cache racing is having multiple tasks wanting to cache the same file
when it currently isn't present in memory.

This was caused by `maybe_cache_file` only checking if the file was
currently in cache, but not if it was already being in the process of
doing so.

This patch adds a synchronisation between cache tasks, preventing the
`maybe_cache_function` from caching the file at the same time.