[cgi] Add infrastructure to support CGI processing
CGI is the Common Gateway Interface, a way to forward requests to
scripts (e.g. bash, perl, etc.) which can handle the requests and
respond to them. This is done by creating a new process and sending the
request to the scripts via stdout, and reading headers and body from stdout.