[servente] Fix errors on inconsistent cURL feature support
The cURL exit code is now checked based on the status that it indicates,
for example `curl --compressed` is not available on Windows, but that
doesn't mean servente is at fault here. In the future, tests should also
exists that run the servente binary and test the common client browsers,
something using WebDriver.

<https://curl.se/docs/manpage.html#0>